Create autorun helper for starting and stopping the server
