+++ lib/describeApi.js
@@ -4,6 +4,8 @@ const request = require('supertest');
 const chai = require('chai');
 const chaiSubset = require('chai-subset');
 
+const autorun = require('./autorun');
+
 chai.use(chaiSubset);
 
 const usingJest = !!(global.beforeAll);
@@ -284,5 +286,7 @@ function isRawSequalize(obj = {}) {
 	const item = Array.isArray(obj.data) ? obj.data[0] : obj.data;
 	return (typeof item === 'object' && item.sequelize);
 }
+
+describeApi.autorun = autorun;
 
 module.exports = describeApi;
